Rework user endpoints to be a little smarter.
The user endpoints have no data validation against the database to
ensure a user exists before the db operation is attempted. This change
corrects that behavior and moves the burdon of validation from the db
model to the endpoints.

+++ server.js
@@ -34,20 +34,32 @@ server.post("/users", (req, res) => {
 });
 
 server.put("/users/:id", (req, res) => {
-   const user = db.updateUser(req.params.id, {
-      name: req.body.name,
-   });
+   const id = req.params.id;
+   const user = db.getUserById(id);
 
    if (user) {
+      db.updateUser(req.params.id, {
+         name: req.body.name,
+      });
+
       res.json(user);
+      return;
    }
 
    res.status(404).json({ message: "User not found" });
 });
 
 server.delete("/users/:id", (req, res) => {
-   db.deleteUser(req.params.id);
-   res.json({ message: "User removed" });
+   const id = req.params.id;
+   const user = db.getUserById(id);
+
+   if (user) {
+      db.deleteUser(req.params.id);
+      res.json({ message: "User removed" });
+      return;
+   }
+
+   res.status(404).json({ message: "User not found" });
 });
 
 server.listen(8080, () => {
